update(structure) : as recommended in the review , I modulized the class WeekProgDailyObj and breaked into :
WeekProgressSection,DailyObjSection,TestTags,WeekDotRow,GlassSurface(to make it usable across the project)
